Separate into cubing module.
